+++ packages/tools/client-debugger/client-debugger-view/src/ContainerHistoryLog.tsx
@@ -1,0 +1,126 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import React from "react";
+import {
+	tokens,
+	TableBody,
+	TableCell,
+	TableRow,
+	Table,
+	TableHeader,
+	TableHeaderCell,
+} from "@fluentui/react-components";
+import {
+	Clock20Regular,
+	PlugConnected24Regular,
+	Attach24Regular,
+	PlugDisconnected24Regular,
+	ErrorCircle24Regular,
+	Warning24Regular,
+	ClipboardError24Filled,
+} from "@fluentui/react-icons";
+import { ConnectionStateChangeLogEntry } from "@fluid-tools/client-debugger";
+
+/**
+ * Represents container state history data which is rendered in {@link ContainerHistoryLog}.
+ */
+export interface ContainerHistoryLogProps {
+	/**
+	 * containerHistory containing clientId & StateChangeLogEntry data.
+	 */
+	containerHistory: readonly ConnectionStateChangeLogEntry[] | undefined;
+}
+
+/**
+ * Renders current state of the connected container.
+ */
+export function ContainerHistoryLog(props: ContainerHistoryLogProps): React.ReactElement {
+	const { containerHistory } = props;
+
+	// Columns for rendering container state history.
+	const containerHistoryColumns = [
+		{ columnKey: "state", label: "State" },
+		{ columnKey: "time", label: "Time" },
+	];
+
+	const getBackgroundColorForState = (state: string): string => {
+		switch (state) {
+			case "connected":
+				return tokens.colorPaletteGreenBackground2; // green
+			case "disconnected":
+				return tokens.colorPaletteDarkOrangeBorderActive; // orange
+			case "closed":
+				return tokens.colorPaletteRedBorder1; // red
+			case "disposed":
+				return tokens.colorPaletteDarkRedBackground2; // dark red
+			case "attached":
+				return tokens.colorPaletteRoyalBlueBackground2; // blue
+			default:
+				console.log("Unknown state type for container!");
+				return tokens.colorBrandBackgroundPressed; // black
+		}
+	};
+
+	return (
+		<Table size="small" aria-label="Audience history table">
+			<TableHeader>
+				<TableRow>
+					{containerHistoryColumns.map((column, columnIndex) => (
+						<TableHeaderCell key={columnIndex}>
+							{column.columnKey === "state" && <PlugConnected24Regular />}
+							{column.columnKey === "time" && <Clock20Regular />}
+							{column.label}
+						</TableHeaderCell>
+					))}
+				</TableRow>
+			</TableHeader>
+			<TableBody>
+				{containerHistory?.map((item, itemIndex) => {
+					const nowTimeStamp = new Date();
+					const changeTimeStamp = new Date(item.timestamp);
+					const wasChangeToday = nowTimeStamp.getDate() === changeTimeStamp.getDate();
+
+					const timestampDisplay = wasChangeToday
+						? changeTimeStamp.toTimeString()
+						: changeTimeStamp.toDateString();
+
+					const getStateIcon = (state: string): React.ReactElement => {
+						switch (state) {
+							case "connected":
+								return <PlugConnected24Regular />;
+							case "attached":
+								return <Attach24Regular />;
+							case "disconnected":
+								return <PlugDisconnected24Regular />;
+							case "disposed":
+								return <ErrorCircle24Regular />;
+							case "closed":
+								return <ClipboardError24Filled />;
+							default:
+								console.log("Unknown state type for container!");
+								return <Warning24Regular />;
+						}
+					};
+
+					return (
+						<TableRow
+							key={itemIndex}
+							style={{
+								backgroundColor: getBackgroundColorForState(item.newState),
+							}}
+						>
+							<TableCell>
+								{getStateIcon(item.newState)}
+								{item.newState}
+							</TableCell>
+							<TableCell>{timestampDisplay}</TableCell>
+						</TableRow>
+					);
+				})}
+			</TableBody>
+		</Table>
+	);
+}
